daemonize moved from gu-persist to gu-base
