+++ gu-base/Cargo.toml
@@ -18,3 +18,5 @@ futures-cpupool = "0.1.8"
 sha1 = "0.6.0"
 tar = "0.4.18"
 flate2 = "1.0.4"
+daemonize = "0.3.0"
+libc = "0.2.43"
