+++ gu-base/src/lib.rs
@@ -13,6 +13,14 @@ extern crate tar;
 extern crate lazy_static;
 extern crate env_logger;
 
+extern crate daemonize;
+extern crate libc;
+
+#[cfg(unix)]
+mod daemon;
+#[cfg(unix)]
+pub mod daemon_module;
+
 pub mod cli;
 pub mod files;
 mod output;
