+++ gu-persist/Cargo.toml
@@ -10,8 +10,6 @@ serde_json = "1.0"
 error-chain = "0.12"
 futures = "0.1"
 directories = "1.0"
-daemonize = "0.3.0"
-libc = "0.2.43"
 bytes = "0.4.10"
 
 actix = "0.7"
